+++ vast_pipeline/views.py
@@ -1333,6 +1333,9 @@ def SourceDetail(request, pk):
             frequency=F('image__band__frequency'),
         ).order_by('datetime').values(*tuple(cols))
     )
+
+    first_det_meas_index = [i['forced'] for i in measurements].index(False)
+
     for one_m in measurements:
         one_m['datetime'] = one_m['datetime'].isoformat()
 
@@ -1444,6 +1447,7 @@ def SourceDetail(request, pk):
         'source': source,
         'source_next_id': source_next_id,
         'source_previous_id': source_previous_id,
+        'first_det_meas_index': first_det_meas_index,
         'datatables': [measurements, related_datatables],
         # flag to deactivate starring and render yellow star
         'sourcefav': (
@@ -2062,10 +2066,10 @@ class UtilitiesSet(ViewSet):
             simbad_results_df = simbad_result_table[
                 ["MAIN_ID", "DISTANCE_RESULT", "OTYPE_S", "OTYPE_V", "RA", "DEC"]
             ].to_pandas()
-            bytestring_fields = ["MAIN_ID", "OTYPE_S", "OTYPE_V"]
-            simbad_results_df[bytestring_fields] = simbad_results_df[
-                bytestring_fields
-            ].apply(lambda col: col.str.decode("utf-8"))
+            # bytestring_fields = ["MAIN_ID", "OTYPE_S", "OTYPE_V"]
+            # simbad_results_df[bytestring_fields] = simbad_results_df[
+            #     bytestring_fields
+            # ].apply(lambda col: col.str.decode("utf-8"))
             simbad_results_df = simbad_results_df.rename(
                 columns={
                     "MAIN_ID": "object_name",
@@ -2094,10 +2098,10 @@ class UtilitiesSet(ViewSet):
             ned_results_df = ned_result_table[
                 ["Object Name", "Separation", "Type", "RA", "DEC"]
             ].to_pandas()
-            bytestring_fields = ["Object Name", "Type"]
-            ned_results_df[bytestring_fields] = ned_results_df[bytestring_fields].apply(
-                lambda col: col.str.decode("utf-8")
-            )
+            # bytestring_fields = ["Object Name", "Type"]
+            # ned_results_df[bytestring_fields] = ned_results_df[bytestring_fields].apply(
+            #     lambda col: col.str.decode("utf-8")
+            # )
             ned_results_df = ned_results_df.rename(
                 columns={
                     "Object Name": "object_name",